Update Anu/Hmwk1/hmwk.py
Well-commented, but no objects and lots of duplicated code.
There are a few errors that result in the wrong answers, but I think that could be helped a lot
if you package the code as objects and separate out all the operations
into logical methods-- keep practicing! And consider reading about/taking a course in OO programming.