Update James/btcmp_hmwk1.py
Well done. Clean and readable code!
For loops can be limited with a few special array functions (see my solution for some examples),
but that comes with time and experience, and wasn't expected.
Too bad you didn't turn it in on time-- would have been a contender for best code sample!